backend/requirements.txt
fastapi
uvicorn
dotenv
openai
google-genai
sqlalchemy
passlib
python-jose
